#23 - FeatureEditor and FeatureSupport for Modifiers of a fact
- move files to ui-kb-feature
- move the style to css
- remove unnecessary dependency
- refactor property/subjectObject feature supports
