+++ inception-ui-kb/src/main/java/de/tudarmstadt/ukp/inception/ui/kb/feature/SubjectObjectFeatureEditor.html
@@ -1,21 +1,21 @@
 <!DOCTYPE html>
 <!--
- - Copyright 2018
- - Ubiquitous Knowledge Processing (UKP) Lab
- - Technische Universität Darmstadt
- -
- - Licensed under the Apache License, Version 2.0 (the "License");
- - you may not use this file except in compliance with the License.
- - You may obtain a copy of the License at
- -
- - http://www.apache.org/licenses/LICENSE-2.0
- -
- - Unless required by applicable law or agreed to in writing, software
- - distributed under the License is distributed on an "AS IS" BASIS,
- - WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- - See the License for the specific language governing permissions and
- - limitations under the License.
--->
+  ~ Copyright 2018
+  ~ Ubiquitous Knowledge Processing (UKP) Lab
+  ~ Technische Universität Darmstadt
+  ~
+  ~ Licensed under the Apache License, Version 2.0 (the "License");
+  ~ you may not use this file except in compliance with the License.
+  ~ You may obtain a copy of the License at
+  ~
+  ~ http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS,
+  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~ See the License for the specific language governing permissions and
+  ~ limitations under the License.
+  -->
 
 <html xmlns="http://www.w3.org/1999/xhtml"
     xmlns:wicket="http://wicket.apache.org/dtds.data/wicket-xhtml1.4-strict.dtd">
@@ -28,23 +28,16 @@
     </div>
     <div class="panel-body">
       <div wicket:id="content">
-        <div class="form-group">
-          <div class="col-sm-12">
-            <div class="input-group">
-              <span wicket:id="label" class="form-control text-nowrap feature-editor-slot-label"></span>
-              <span class="input-group-btn">
-                <button wicket:id="removeLabel" type="button" class="btn btn-default">
-                  <i class="fa fa-times" aria-hidden="true" aria-label="clear"></i>
-                </button>
+          <div class="form-group">
+            <label class="col-sm-3 control-label">
+              <span wicket:id="role"></span>
+            </label>
+            <div class="col-sm-9">
+              <span wicket:id="label" class="form-control text-nowrap feature-editor-slot-label">
               </span>
             </div>
           </div>
-        </div>
-        <div class="form-group">
-          <div class="col-sm-12">
-            <select style="width: 100%;" wicket:id="value"></select>
-          </div>
-        </div>
+          <td><select style="width: 100%;" wicket:id="value"></select></td>
       </div>
     </div>
   </div>
