+++ inception-ui-kb/src/main/java/de/tudarmstadt/ukp/inception/ui/kb/feature/SubjectObjectFeatureEditor.html~HEAD
@@ -28,16 +28,23 @@
     </div>
     <div class="panel-body">
       <div wicket:id="content">
-          <div class="form-group">
-            <label class="col-sm-3 control-label">
-              <span wicket:id="role"></span>
-            </label>
-            <div class="col-sm-9">
-              <span wicket:id="label" class="form-control text-nowrap"
-                  style="max-width: 20em; overflow:hidden; text-overflow: ellipsis;"></span>
+        <div class="form-group">
+          <div class="col-sm-12">
+            <div class="input-group">
+              <span wicket:id="label" class="form-control text-nowrap feature-editor-slot-label"></span>
+              <span class="input-group-btn">
+                <button wicket:id="removeLabel" type="button" class="btn btn-default">
+                  <i class="fa fa-times" aria-hidden="true" aria-label="clear"></i>
+                </button>
+              </span>
             </div>
           </div>
-          <td><select style="width: 100%;" wicket:id="value"></select></td>
+        </div>
+        <div class="form-group">
+          <div class="col-sm-12">
+            <select style="width: 100%;" wicket:id="value"></select>
+          </div>
+        </div>
       </div>
     </div>
   </div>
