+++ inception-ui-kb/src/main/java/de/tudarmstadt/ukp/inception/ui/kb/feature/SubjectObjectFeatureEditor.java
@@ -22,8 +22,6 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.NoSuchElementException;
 
-import org.apache.commons.lang3.exception.ExceptionUtils;
-import org.apache.uima.UIMAException;
 import org.apache.uima.cas.CASException;
 import org.apache.uima.cas.Feature;
 import org.apache.uima.cas.text.AnnotationFS;
@@ -46,7 +44,6 @@ import org.slf4j.LoggerFactory;
 import com.googlecode.wicket.kendo.ui.form.dropdown.DropDownList;
 import de.tudarmstadt.ukp.clarin.webanno.api.AnnotationSchemaService;
 import de.tudarmstadt.ukp.clarin.webanno.api.annotation.action.AnnotationActionHandler;
-import de.tudarmstadt.ukp.clarin.webanno.api.annotation.exception.AnnotationException;
 import de.tudarmstadt.ukp.clarin.webanno.api.annotation.feature.editor.FeatureEditor;
 import de.tudarmstadt.ukp.clarin.webanno.api.annotation.model.AnnotatorState;
 import de.tudarmstadt.ukp.clarin.webanno.api.annotation.model.FeatureState;
@@ -54,28 +51,33 @@ import de.tudarmstadt.ukp.clarin.webanno.api.annotation.model.LinkWithRoleModel;
 import de.tudarmstadt.ukp.clarin.webanno.api.annotation.util.WebAnnoCasUtil;
 import de.tudarmstadt.ukp.clarin.webanno.model.AnnotationFeature;
 import de.tudarmstadt.ukp.clarin.webanno.model.AnnotationLayer;
-import de.tudarmstadt.ukp.clarin.webanno.support.lambda.LambdaAjaxLink;
 import de.tudarmstadt.ukp.clarin.webanno.support.lambda.LambdaModel;
 import de.tudarmstadt.ukp.clarin.webanno.support.lambda.LambdaModelAdapter;
 import de.tudarmstadt.ukp.inception.kb.KnowledgeBaseService;
 import de.tudarmstadt.ukp.inception.kb.graph.KBHandle;
+import de.tudarmstadt.ukp.inception.kb.graph.KBStatement;
 import de.tudarmstadt.ukp.inception.kb.model.KnowledgeBase;
 
 public class SubjectObjectFeatureEditor
     extends FeatureEditor
 {
+
     private static final long serialVersionUID = 4230722501745589589L;
-    private static final Logger LOG = LoggerFactory.getLogger(SubjectObjectFeatureEditor.class);
-
+    private static final Logger logger = LoggerFactory.getLogger(SubjectObjectFeatureEditor.class);
     private @SpringBean AnnotationSchemaService annotationService;
-    private @SpringBean KnowledgeBaseService kbService;
-
     private WebMarkupContainer content;
-    private Component focusComponent;
+
+    @SuppressWarnings("rawtypes") private Component focusComponent;
+    private boolean hideUnconstraintFeature;
+
     private AnnotationActionHandler actionHandler;
     private IModel<AnnotatorState> stateModel;
-    private LinkWithRoleModel roleModel;
-    private AnnotationFeature linkedAnnotationFeature;
+
+    @SuppressWarnings("unused") private LinkWithRoleModel roleModel;
+
+    AnnotationFeature linkedAnnotationFeature;
+
+    private @SpringBean KnowledgeBaseService kbService;
 
     public SubjectObjectFeatureEditor(String aId, MarkupContainer aOwner,
         AnnotationActionHandler aHandler, final IModel<AnnotatorState> aStateModel,
@@ -85,6 +87,8 @@ public class SubjectObjectFeatureEditor
 
         stateModel = aStateModel;
         actionHandler = aHandler;
+
+        hideUnconstraintFeature = getModelObject().feature.isHideUnconstraintFeature();
 
         add(new Label("feature", getModelObject().feature.getUiName()));
         content = new WebMarkupContainer("content");
@@ -100,13 +104,14 @@ public class SubjectObjectFeatureEditor
             roleModel = links.get(0);
         }
 
+        content.add(new Label("role", roleModel.role));
         content.add(createSubjectObjectLabel());
-        content.add(createRemoveLabelIcon());
         content.add(focusComponent = createFieldComboBox());
     }
 
     private Label createSubjectObjectLabel()
     {
+        AnnotatorState state = stateModel.getObject();
         Label label;
         label = new Label("label", LambdaModel.of(this::getSelectionSlotLabel));
         label.add(new AjaxEventBehavior("click")
@@ -137,32 +142,6 @@ public class SubjectObjectFeatureEditor
             label.setDefaultModelObject(roleModel.label);
         }
         return label;
-    }
-
-    private LambdaAjaxLink createRemoveLabelIcon()
-    {
-        return new LambdaAjaxLink("removeLabel", this::removeSelectedLabel);
-    }
-
-    private void removeSelectedLabel(AjaxRequestTarget aTarget)
-    {
-        List<LinkWithRoleModel> links = (List<LinkWithRoleModel>) this.getModelObject().value;
-        AnnotatorState state = this.stateModel.getObject();
-
-        String role = roleModel.role;
-        roleModel = new LinkWithRoleModel();
-        roleModel.role = role;
-        links.set(0, roleModel);
-
-        // Auto-commit if working on existing annotation
-        if (state.getSelection().getAnnotation().isSet()) {
-            try {
-                actionHandler.actionCreateOrUpdate(aTarget, actionHandler.getEditorCas());
-            }
-            catch (Exception e) {
-                handleException(this, aTarget, e);
-            }
-        }
     }
 
     private DropDownList<KBHandle> createFieldComboBox()
@@ -201,11 +180,11 @@ public class SubjectObjectFeatureEditor
 
         if (roleLabelSlotIsSelected()) {
             state.clearArmedSlot();
-            aTarget.add(getOwner());
+            aTarget.add(content);
         }
         else {
             state.setArmedSlot(getModelObject().feature, 0);
-            aTarget.add(getOwner());
+            aTarget.add(content);
         }
     }
 
@@ -252,8 +231,12 @@ public class SubjectObjectFeatureEditor
                     .setFeature(selectedFS, linkedAnnotationFeature, value.getIdentifier());
             }
             catch (CASException | IOException e) {
-                LOG.error("Error: " + e.getMessage(), e);
+                logger.error("Error: " + e.getMessage(), e);
                 error("Error: " + e.getMessage());
+            }
+            if (roleModel.role.equals("subject")) {
+                setStatementInKB(value);
+
             }
         }
     }
@@ -265,9 +248,9 @@ public class SubjectObjectFeatureEditor
             try {
                 JCas jCas = actionHandler.getEditorCas().getCas().getJCas();
                 AnnotationFS selectedFS = WebAnnoCasUtil.selectByAddr(jCas, roleModel.targetAddr);
-                Feature labelFeature = selectedFS.getType()
-                    .getFeatureByBaseName(linkedAnnotationFeature.getName());
-                String selectedKBItemIdentifier = selectedFS.getFeatureValueAsString(labelFeature);
+                String selectedKBItemIdentifier = WebAnnoCasUtil.getFeature(selectedFS,
+                    linkedAnnotationFeature.getName());
+
                 if (selectedKBItemIdentifier != null) {
                     List<KBHandle> handles = getKBConceptsAndInstances();
                     selectedKBHandleItem = handles.stream()
@@ -276,7 +259,7 @@ public class SubjectObjectFeatureEditor
                 }
             }
             catch (CASException | IOException e) {
-                LOG.error("Error: " + e.getMessage(), e);
+                logger.error("Error: " + e.getMessage(), e);
                 error("Error: " + e.getMessage());
             }
         }
@@ -296,28 +279,71 @@ public class SubjectObjectFeatureEditor
         return handles;
     }
 
-    public static void handleException(Component aComponent, AjaxRequestTarget aTarget,
-        Exception aException)
-    {
-        try {
-            throw aException;
-        }
-        catch (AnnotationException e) {
-            if (aTarget != null) {
-                aTarget.prependJavaScript("alert('Error: " + e.getMessage() + "')");
+    private KnowledgeBase getKBForKBHandle(KBHandle kbHandle) {
+        AnnotationFeature feat = getModelObject().feature;
+        for (KnowledgeBase kb: kbService.getKnowledgeBases(feat.getProject())) {
+            if (kbService.listProperties(kb, false).contains(kbHandle)) {
+                return kb;
             }
-            else {
-                aComponent.error("Error: " + e.getMessage());
+            if (kbService.listConcepts(kb, false).contains(kbHandle)) {
+                return kb;
             }
-            LOG.error("Error: " + ExceptionUtils.getRootCauseMessage(e), e);
+            for (KBHandle concept: kbService.listConcepts(kb, false)) {
+                if (kbService.listInstances(kb, concept.getIdentifier(), false).contains
+                    (kbHandle)) {
+                    return kb;
+                }
+            }
+
         }
-        catch (UIMAException e) {
-            aComponent.error("Error: " + ExceptionUtils.getRootCauseMessage(e));
-            LOG.error("Error: " + ExceptionUtils.getRootCauseMessage(e), e);
-        }
-        catch (Exception e) {
-            aComponent.error("Error: " + e.getMessage());
-            LOG.error("Error: " + e.getMessage(), e);
+        return null;
+    }
+
+    private void setStatementInKB(KBHandle value) {
+        AnnotationLayer factLayer = annotationService.getLayer(
+            "webanno.custom.Fact", this.stateModel.getObject().getProject());
+        AnnotationFeature predicateFeature = annotationService.getFeature(
+            "KBPredicate", factLayer);
+        KBHandle predicateHandle = (KBHandle) this.stateModel.getObject().getFeatureState
+            (predicateFeature).value;
+        if (predicateHandle != null) {
+            KBStatement statement = new KBStatement();
+            statement.setInstance(value);
+            statement.setProperty(predicateHandle);
+            AnnotationFeature objectFeature = annotationService.getFeature("Object",
+                factLayer);
+            List<LinkWithRoleModel> objectList = (List<LinkWithRoleModel>) this.stateModel.
+                getObject().getFeatureState(objectFeature).value;
+//            String objectLabel = objectList.get(0).label;
+//            String object = objectLabel.equals("<Click to activate>") || objectLabel.equals
+//                ("<Select to fill>") ? " " : objectLabel;
+            int targetAddress = objectList.get(0).targetAddr;
+            String object = "";
+            if (targetAddress != -1) {
+                try {
+                    JCas jCas = actionHandler.getEditorCas().getCas().getJCas();
+                    AnnotationFS selectedFS = WebAnnoCasUtil.selectByAddr(jCas, targetAddress);
+                    String selectedKBItemIdentifier = WebAnnoCasUtil.getFeature(selectedFS,
+                        linkedAnnotationFeature.getName());
+                    if (selectedKBItemIdentifier != null) {
+                        List<KBHandle> handles = getKBConceptsAndInstances();
+                        object = handles.stream().filter(x -> selectedKBItemIdentifier
+                            .equals(x.getIdentifier())).findAny().orElse(null).getUiLabel();
+                    }
+                } catch (CASException | IOException e) {
+                    logger.error("Error: " + e.getMessage(), e);
+                    error("Error: " + e.getMessage());
+                }
+            }
+            statement.setValue(object);
+            KnowledgeBase subjectKB = getKBForKBHandle(value);
+            KnowledgeBase predicateKB = getKBForKBHandle(predicateHandle);
+            if (subjectKB.equals(predicateKB)) {
+                kbService.upsertStatement(subjectKB, statement);
+            } else {
+                logger.error("Subject and predicate are from different knowledge bases.");
+            }
         }
     }
+
 }
