+++ inception-ui-kb/src/main/java/de/tudarmstadt/ukp/inception/ui/kb/feature/SubjectObjectFeatureSupport.java
@@ -30,6 +30,8 @@ import org.apache.uima.resource.metadata.TypeDescription;
 import org.apache.uima.resource.metadata.TypeSystemDescription;
 import org.apache.wicket.MarkupContainer;
 import org.apache.wicket.model.IModel;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.core.Ordered;
 import org.springframework.core.annotation.Order;
 import org.springframework.stereotype.Component;
@@ -51,8 +53,10 @@ import de.tudarmstadt.ukp.clarin.webanno.model.AnnotationLayer;
 @Order(Ordered.HIGHEST_PRECEDENCE)
 @Component
 public class SubjectObjectFeatureSupport
-    implements FeatureSupport<Void>
+    implements FeatureSupport
 {
+    private final Logger logger = LoggerFactory.getLogger(getClass());
+
     private static final String SUBJECT_LINK = "webanno.custom.FactSubjectLink";
     private static final String OBJECT_LINK = "webanno.custom.FactObjectLink";
     private static final String MODIFIER_LINK = "webanno.custom.FactModifierLink";
@@ -87,8 +91,8 @@ public class SubjectObjectFeatureSupport
             switch (annotationFeature.getLinkMode()) {
             case WITH_ROLE:
                 switch (annotationFeature.getLinkTypeName()) {
-                case SUBJECT_LINK: // fall-through
-                case OBJECT_LINK: // fall-through
+                case SUBJECT_LINK: // fallthrough
+                case OBJECT_LINK: // fallthrough
                 case MODIFIER_LINK:
                     return true;
                 default:
@@ -97,7 +101,7 @@ public class SubjectObjectFeatureSupport
             default:
                 return false;
             }
-        case NONE: // fall-through
+        case NONE: // fallthrough
         default:
             return false;
         }
@@ -125,22 +129,22 @@ public class SubjectObjectFeatureSupport
                     editor = new SubjectObjectFeatureEditor(aId, aOwner, aHandler, aStateModel,
                         aFeatureStateModel, OBJECT_ROLE);
                     break;
-//                case MODIFIER_LINK:
-//                    editor = new ModifierFeatureEditor(aId, aOwner, aHandler, aStateModel,
-//                        aFeatureStateModel);
-//                    break;
+                case MODIFIER_LINK:
+                    editor = new ModifierFeatureEditor(aId, aOwner, aHandler, aStateModel,
+                        aFeatureStateModel);
+                    break;
                 default:
-                    throw unsupportedLinkModeException(featureState.feature);
+                    throw unsupportedLinkModeException(featureState);
                 }
                 break;
             default:
-                throw unsupportedMultiValueModeException(featureState.feature);
+                throw unsupportedMultiValueModeException(featureState);
             }
             break;
         case NONE:
-            throw unsupportedFeatureTypeException(featureState.feature);
+            throw unsupportedFeatureTypeException(featureState);
         default:
-            throw unsupportedMultiValueModeException(featureState.feature);
+            throw unsupportedMultiValueModeException(featureState);
         }
         return editor;
     }
@@ -154,7 +158,6 @@ public class SubjectObjectFeatureSupport
             CAS.TYPE_NAME_TOP);
         linkTD.addFeature(aFeature.getLinkTypeRoleFeatureName(), "", CAS.TYPE_NAME_STRING);
         linkTD.addFeature(aFeature.getLinkTypeTargetFeatureName(), "", aFeature.getType());
-        
         // Link feature
         aTD.addFeature(aFeature.getName(), "", CAS.TYPE_NAME_FS_ARRAY, linkTD.getName(),
             false);
