+++ inception-ui-kb/src/main/java/de/tudarmstadt/ukp/inception/ui/kb/feature/SubjectObjectFeatureSupport.java~HEAD
@@ -7,7 +7,7 @@
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *  http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -15,15 +15,21 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.tudarmstadt.ukp.inception.kb.feature;
+package de.tudarmstadt.ukp.inception.ui.kb.feature;
 
 import java.util.ArrayList;
 import java.util.List;
 
+import org.apache.uima.cas.ArrayFS;
+import org.apache.uima.cas.CAS;
+import org.apache.uima.cas.Feature;
+import org.apache.uima.cas.FeatureStructure;
+import org.apache.uima.cas.Type;
+import org.apache.uima.cas.text.AnnotationFS;
+import org.apache.uima.resource.metadata.TypeDescription;
+import org.apache.uima.resource.metadata.TypeSystemDescription;
 import org.apache.wicket.MarkupContainer;
 import org.apache.wicket.model.IModel;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.core.Ordered;
 import org.springframework.core.annotation.Order;
 import org.springframework.stereotype.Component;
@@ -34,6 +40,8 @@ import de.tudarmstadt.ukp.clarin.webanno.api.annotation.feature.FeatureType;
 import de.tudarmstadt.ukp.clarin.webanno.api.annotation.feature.editor.FeatureEditor;
 import de.tudarmstadt.ukp.clarin.webanno.api.annotation.model.AnnotatorState;
 import de.tudarmstadt.ukp.clarin.webanno.api.annotation.model.FeatureState;
+import de.tudarmstadt.ukp.clarin.webanno.api.annotation.model.LinkWithRoleModel;
+import de.tudarmstadt.ukp.clarin.webanno.api.annotation.util.WebAnnoCasUtil;
 import de.tudarmstadt.ukp.clarin.webanno.model.AnnotationFeature;
 import de.tudarmstadt.ukp.clarin.webanno.model.AnnotationLayer;
 
@@ -43,15 +51,27 @@ import de.tudarmstadt.ukp.clarin.webanno.model.AnnotationLayer;
 @Order(Ordered.HIGHEST_PRECEDENCE)
 @Component
 public class SubjectObjectFeatureSupport
-    implements FeatureSupport
+    implements FeatureSupport<Void>
 {
-    private final Logger logger = LoggerFactory.getLogger(getClass());
-
     private static final String SUBJECT_LINK = "webanno.custom.FactSubjectLink";
     private static final String OBJECT_LINK = "webanno.custom.FactObjectLink";
     private static final String MODIFIER_LINK = "webanno.custom.FactModifierLink";
     private static final String SUBJECT_ROLE = "subject";
     private static final String OBJECT_ROLE = "object";
+
+    private String featureSupportId;
+
+    @Override
+    public String getId()
+    {
+        return featureSupportId;
+    }
+
+    @Override
+    public void setBeanName(String aBeanName)
+    {
+        featureSupportId = aBeanName;
+    }
 
     @Override
     public List<FeatureType> getSupportedFeatureTypes(AnnotationLayer aAnnotationLayer)
@@ -67,8 +87,8 @@ public class SubjectObjectFeatureSupport
             switch (annotationFeature.getLinkMode()) {
             case WITH_ROLE:
                 switch (annotationFeature.getLinkTypeName()) {
-                case SUBJECT_LINK: // fallthrough
-                case OBJECT_LINK: // fallthrough
+                case SUBJECT_LINK: // fall-through
+                case OBJECT_LINK: // fall-through
                 case MODIFIER_LINK:
                     return true;
                 default:
@@ -77,7 +97,7 @@ public class SubjectObjectFeatureSupport
             default:
                 return false;
             }
-        case NONE: // fallthrough
+        case NONE: // fall-through
         default:
             return false;
         }
@@ -105,24 +125,65 @@ public class SubjectObjectFeatureSupport
                     editor = new SubjectObjectFeatureEditor(aId, aOwner, aHandler, aStateModel,
                         aFeatureStateModel, OBJECT_ROLE);
                     break;
-                case MODIFIER_LINK:
-                    editor = new ModifierFeatureEditor(aId, aOwner, aHandler, aStateModel,
-                        aFeatureStateModel);
-                    break;
+//                case MODIFIER_LINK:
+//                    editor = new ModifierFeatureEditor(aId, aOwner, aHandler, aStateModel,
+//                        aFeatureStateModel);
+//                    break;
                 default:
-                    throw unsupportedLinkModeException(featureState);
+                    throw unsupportedLinkModeException(featureState.feature);
                 }
                 break;
             default:
-                throw unsupportedMultiValueModeException(featureState);
+                throw unsupportedMultiValueModeException(featureState.feature);
             }
             break;
         case NONE:
-            throw unsupportedFeatureTypeException(featureState);
+            throw unsupportedFeatureTypeException(featureState.feature);
         default:
-            throw unsupportedMultiValueModeException(featureState);
+            throw unsupportedMultiValueModeException(featureState.feature);
         }
         return editor;
+    }
+
+    @Override
+    public void generateFeature(TypeSystemDescription aTSD, TypeDescription aTD,
+        AnnotationFeature aFeature)
+    {
+        // Link type
+        TypeDescription linkTD = aTSD.addType(aFeature.getLinkTypeName(), "",
+            CAS.TYPE_NAME_TOP);
+        linkTD.addFeature(aFeature.getLinkTypeRoleFeatureName(), "", CAS.TYPE_NAME_STRING);
+        linkTD.addFeature(aFeature.getLinkTypeTargetFeatureName(), "", aFeature.getType());
+        
+        // Link feature
+        aTD.addFeature(aFeature.getName(), "", CAS.TYPE_NAME_FS_ARRAY, linkTD.getName(),
+            false);
+    }
+
+    @Override
+    public <T> T getFeatureValue(AnnotationFeature aFeature, FeatureStructure aFS)
+    {
+        // Get type and features - we need them later in the loop
+        Feature linkFeature = aFS.getType().getFeatureByBaseName(aFeature.getName());
+        Type linkType = aFS.getCAS().getTypeSystem().getType(aFeature.getLinkTypeName());
+        Feature roleFeat = linkType.getFeatureByBaseName(aFeature
+            .getLinkTypeRoleFeatureName());
+        Feature targetFeat = linkType.getFeatureByBaseName(aFeature
+            .getLinkTypeTargetFeatureName());
+
+        List<LinkWithRoleModel> links = new ArrayList<>();
+        ArrayFS array = (ArrayFS) aFS.getFeatureValue(linkFeature);
+        if (array != null) {
+            for (FeatureStructure link : array.toArray()) {
+                LinkWithRoleModel m = new LinkWithRoleModel();
+                m.role = link.getStringValue(roleFeat);
+                m.targetAddr = WebAnnoCasUtil.getAddr(link.getFeatureValue(targetFeat));
+                m.label = ((AnnotationFS) link.getFeatureValue(targetFeat))
+                    .getCoveredText();
+                links.add(m);
+            }
+        }
+        return (T) links;
     }
 
 }
